Close DatasetReader in batch-processing.py
Hey 👋, I'm trying this change out myself to see if it fixes the memory leak I mentioned earlier today on Discord. So far so good on my end but I want to let it ride for a little longer for good measure. I'll report back but want to run it by you first to see if you think a change like this makes sense to you or if you think the change is not worth it or doesn't make sense for some reason 😊